Some ideas after coming across report generation failures in jasper report server with scheduled reports. The issue could possibly be addressed in the usage of the virtualizer, since pageOut is being called after dispose. However some fixes can be done in the shared swap virtualizer system to address and fix the issue, and make it generally more thread safe regardless of how it is used.
The StoreFactoryVirtualizerTest is added to reproduce various issues in
concurrent use of the swap file virtualizer, from concurrent pageOuts to
concurrent pageOut/pageIn, to dispose being called prematurely, to
concurrent dispose calls within pageOut and pageIn calls.

A optional new flag is proposed that can be set to not dispose if
handles are still present in the store. Default existing behavior is
unchanged.

The interfaces are unchanged to ensure no other implementations are
broken. This unfortunately leads to casting the store to
'SwapFileVirtualizerStore' in some cases, but this is for backward
compatibility. Although it seems the SwapFileVirtualizerStore is a
strong part of the system. But this could be changed.

Some of the core updates here are
JRSwapFile:
 - syncronization fixes (from unit test cases)
 - isDisposed method

StoreFactoryVirtualizer
 - syncronization fixes around the usage of contextStores (Map)
 - context level lock in pageOut and pageIn like dispose
 - not returning a 'disposed' store
 - removing a disposed store from the map

SwapFileVirtualizerStore
 - syncronization fix around the serializer.write (from unit tests)
 - isDisposed and isSwapFileDisposed methods (disposal of file depends
on ownership)
 - introduce optional 'disposeOnlyWithEmptyHandles' boolean to indicate
to ignore disposal if handles are still present (could possible be
helpful for clients using the api with out of order pageIn calls)

SwapFileVirtualizerStoreFactory
 - add the optional 'disposeOnlyWithEmptyHandles' flag so it can be
configured on the factory and the spring configuration files